handle a case when media length is None in both sources and webui
Not sure about this fix, because I am not sure if None as a media length is a
bug fixed by this patch or a bug in the Xine glue which gives None in some cases
and shouldn't. Well, my problem is fixed for the moment. And I can't find media
in my library which gives a None length.

darcs-hash:20080202175705-85ee7-36fdb3726b0868187f636036d5c09eefe12f6be8.gz

committer: Alexandre Rossi <alexandre.rossi@gmail.com>