[net] fix python oddity uncovered by previous patch (private attr not being private)
A previous patch deleted a constructor that seemed unused because it was
declaring only private variables (with leading '__'). When seing this, python
replaces this with _ClassName__attrname to make it private. Here, the base
class and the derived class have got the same name, and the _ClassName__ prefix
is the same, which means that there was no error.
This is why I was mistaken to think that the attr was private.

darcs-hash:20080609151952-85ee7-7596e79a6c2fda4013096e9786156c1c3b5bd3cb.gz

committer: Alexandre Rossi <alexandre.rossi@gmail.com>