use twisted inotify API
This patch uses the twisted inotify API which runs in the twisted loop
and thus makes useless the thread/Queue mechanism between the inotify thread
and the twisted loop.

This patch also removes special handling of symlinks on directories by
making use of path as seen from the library root instead of the realpath.
This simplifies a lot things.

Finally, this patch fixes longstanding library with inotify bugs :
- directory quick rename and delete which was crashing the inotify
  event handler.
- directory rename
- directory add
- should fix concurrency issues ("Database is locked" errors)